tests/wpt/tests/navigation-api/navigation-methods/return-value/traverseTo-intercept.html

<!doctype html>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="resources/helpers.js"></script>

<script>
promise_test(async t => {
  // Wait for after the load event so that the navigation doesn't get converted
  // into a replace navigation.
  await new Promise(resolve => window.onload = () => t.step_timeout(resolve, 0));

  const key0 = navigation.currentEntry.key;

  location.href = "#1";

  assert_equals(navigation.entries().length, 2);
  const [entry0, entry1] = navigation.entries();
  assert_equals((new URL(entry0.url)).hash, "");
  assert_equals((new URL(entry1.url)).hash, "#1");

  navigation.onnavigate = e => e.intercept({ handler: () => Promise.resolve({ abc: "def" }) });

  const result = navigation.traverseTo(key0);
  await assertBothFulfill(t, result, entry0);
  assert_equals(navigation.currentEntry, entry0);
}, "traverseTo() and intercept() with a fulfilled promise");
</script>
