LayoutTests/imported/w3c/web-platform-tests/navigation-api/navigation-methods/return-value/traverseTo-intercept-expected.txt


FAIL traverseTo() and intercept() with a fulfilled promise assert_equals: expected "#1" but got ""
